refactor(drag-handle-cell): merge drag handle into cell, inline SVG, drop vertical-alignment
