+++ src/assets/styles/settings.css
@@ -730,6 +730,8 @@
 	--semantics-buttons-danger-tinted-is-active-background-color: light-dark(var(--primitives-color-danger-150), var(--primitives-color-danger-250));
 	--semantics-buttons-danger-tinted-is-active-content-color: light-dark(var(--primitives-color-danger-750), var(--primitives-color-danger-800));
 
+	--semantics-grab-handles-background-color: light-dark(var(--primitives-color-neutral-50), var(--primitives-color-neutral-150));
+	--semantics-grab-handles-grip-color: light-dark(var(--primitives-color-neutral-600), var(--primitives-color-neutral-700));
 
 	--semantics-input-fields-border-thickness: var(--primitives-space-2);
 	--semantics-input-fields-placeholder-color: light-dark(var(--primitives-color-neutral-550), var(--primitives-color-neutral-650));
@@ -852,10 +854,6 @@
 
 
 	--components-top-title-bar-button-bar-divider-color: #c9ced8;
-
-
-	--components-grab-handle-background-color: #d9dee4;
-	--components-grab-handle-color: #67778d;
 
 
 	--components-form-gap: var(--primitives-space-16);
